media: dvb: mpq-sdmx: increase number of physical chunks
Buffers used with secure demux may be fragmented into more than
10 physical chunks. In such case current limit will prevent
the filter from starting.
Increase the maximal number of chunks to 256, and shared buffer size
to 128kb in order to accommodate the increased command struct.
As the SDMX application must use the same value, increment the
major version required for API compatibility.

Change-Id: Ie1f3381fa1e391c54978cb3c14305db21e14ef4f
Signed-off-by: Gilad Broner <gbroner@codeaurora.org>